Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-ollama
langchain-community
sentence-transformers
transformers
torch   
chromadb
pdf2image
pytesseract
pillow
reportlab